==> KG_generation/Entity Resolution/1. Player Entity Resolution.ipynb <==
!pip install strsimpy
from strsimpy.jaro_winkler import JaroWinkler
from strsimpy.normalized_levenshtein import NormalizedLevenshtein
import pandas as pd
import numpy as np
import json
import fileinput
from tqdm import tqdm
import string
import re
import json



who_player = pd.DataFrame()
with fileinput.input(files='../' + "whoscored_crawler/player_info_whoscored.jl") as file:
    for line in file:
        conv = json.loads(line)
        who_player = who_player.append(conv, ignore_index=True)
who_player.head()

world_player = pd.DataFrame()
with fileinput.input(files='../' + "worldfootball_crawler/player_info_worldfootball.jl") as file:
    for line in file:
        conv = json.loads(line)
        world_player = world_player.append(conv, ignore_index=True)
world_player.head()

len(who_player)

len(world_player)

                       
len(who_player["player_url"].drop_duplicates())

## ID, Map Generation

len(world_player["player_url"])

len(world_player["player_url"].drop_duplicates())

world_player.groupby("name").count()["player_url"].value_counts()

id2player = dict()
player2id = dict()

for idx, player_url in enumerate(list(world_player["player_url"].drop_duplicates())):
    id2player[idx + 10**6] = player_url
    player2id[player_url] = idx + 10**6

for idx, player_url in enumerate(list(who_player["player_url"].drop_duplicates())):
    id2player[idx + 2*(10**6)] = player_url
    player2id[player_url] = idx + 2 * (10**6)

len(who_player["player_url"])

len(who_player["player_url"].drop_duplicates())

len(id2player)

## Blocking

def f(x):
    if type(x) == float:
        return 0
    elif len(x[0])<15:
        tmp = x[0].split(".")
        return int(str(tmp[2]) + str(tmp[1]) + str(tmp[0]))
    else:
        tmp = x[0].split(".")
        return int(str(tmp[2])[:4] + str(tmp[1]) + str(tmp[0]))
world_player["birth_block"] = world_player["Born"].apply(f)

who_player["DOB"] = who_player["Age"].apply(lambda x : x.split(" years old (")[1][:-1])

who_player["Age"] = who_player["Age"].apply(lambda x : x.split(" years old (")[0])

def f(x):
    tmp = x.split("-")
    return int(str(tmp[2]) + str(tmp[1]) + str(tmp[0]))
who_player["birth_block"] = who_player["DOB"].apply(f)

who_player['player_id'] = who_player["player_url"].apply(lambda x : player2id[x])
world_player['player_id'] = world_player["player_url"].apply(lambda x : player2id[x])

who_birth = who_player[["player_id", "birth_block"]]
world_birth = world_player[["player_id", "birth_block"]]
birth_df = who_birth.append(world_birth)

# birth_df.groupby("birth_block").count().sort_values(by = "player_id", ascending = False)
birth_df

world_name = world_player[["player_id", "name"]]
who_name = who_player[["player_id", "Name"]].rename({"Name" : "name"}, axis =1 )
name_df = world_name.append(who_name)

# Matching

tmp = pd.DataFrame(birth_df.groupby("birth_block")["player_id"].apply(list))
match_dict =dict()
match_dict2 = dict()
id2name = name_df.set_index('player_id').to_dict()["name"]
Lev = NormalizedLevenshtein().similarity

for idx, row in tmp.iterrows():
    if idx != 0 and len(row.player_id)>1:
        name_list_1 = []
        name_list_2 = []
        for idx in row.player_id:
            if str(idx)[0] == "1":
                name_list_1.append((idx,id2name[idx].lower()))
            if str(idx)[0] == "2":
                name_list_2.append((idx,id2name[idx].lower()))
        if len(name_list_1) == 0 or len(name_list_2) == 0:
            continue
        else:
            for (idx1, name1) in name_list_1:
                for (idx2, name2) in name_list_2:
                    if Lev(name1, name2) > 0.4:
#                         if Lev(name1, name2) < 0.5:
#                             print(name1 +"\t" + name2 + "\t" + str(Lev(name1, name2)))
                        try:
                            a = match_dict[idx1]
                            if Lev(name1, name2) > Lev(name1, a[0][1]):
                                match_dict[idx1] = [(idx2, name2)]
                                match_dict2[idx2] = [(idx1, name1)]
                        except:
                            match_dict[idx1] = match_dict.get(idx1, []) + [(idx2, name2)]
                            match_dict2[idx2] = match_dict2.get(idx2, []) + [(idx1, name1)]
        
            
                        
#                         match_dict[idx1] = match_dict.get(idx1, []) + [(idx2, name2)]
#                         match_dict2[idx2] = match_dict2.get(idx2, []) + [(idx1, name1)]
        
        
                        

id_list1 = [i for i in list(id2player.keys()) if str(i)[0] == "1" and i in match_dict.keys()]

id_list2 = [i for i in list(id2player.keys()) if str(i)[0] == "2" and i in match_dict2.keys()]

from collections import Counter

tmp = []
for i in list(match_dict.values()):
    tmp+=i

tmp = []
for i in list(match_dict2.values()):
    tmp+=i

len(id_list1)

len(id_list2)

len(match_dict)

[i for i in match_dict2.items() if len(i[1])!=1]

k = [i for i in match_dict.items() if len(i[1])!=1]
k

def f(x):
    try:
        return match_dict[x][0][0]
    except:
        return "No match"
world_player["match_player_id"] = world_player["player_id"].apply(f)

# outer join in python pandas
outer_join_df=pd.merge(world_player, who_player, left_on = "match_player_id", right_on = "player_id", how='outer')
outer_join_df


outer_join_df.to_csv("../data_" + "player_join.csv")

# PSL DATA Generation


data_dir = 'PSL_data/'

# count = []
# for i in birth_df.groupby("birth_block")["player_id"].apply(list):
#     list1 = []
#     list2 = []
#     if len(i)>1:
#         for ele in i:
#             if str(ele)[0] == '1':
#                 list1.append(ele)
#             else:
#                 list2.append(ele)

#             count.append(len(list1) * len(list2))

# max(count)

# count1 = 0
# count2 = 0
# for i in list(id2player.keys()):
#     if str(i)[0] == '1':
#         count1 += 1
#     else:
#         count2 += 1

## Birth Block
with open(data_dir + "birthBlock_obs.txt", "w") as f:
    for idx, row in birth_df.iterrows():
        f.write(str(row.player_id) + "\t" + str(row.birth_block) + "\n")

# count1*count2

## Name sim
Lev = NormalizedLevenshtein().similarity
with open(data_dir + "simName_obs.txt", "w") as f:
    tmp = pd.DataFrame(birth_df.groupby("birth_block")["player_id"].apply(list))
    for id_list in tmp["player_id"]:
        for idx1 in id_list:
            for idx2 in id_list:
                name1 = str(list(name_df[name_df["player_id"] == int(idx1)]["name"])[0])
                name2 = str(list(name_df[name_df["player_id"] == int(idx2)]["name"])[0])
                f.write(str(idx1) + "\t" + str(idx2) + "\t" + str(Lev(name1.lower(), name2.lower() )) + "\n")

with open(data_dir + "simName_obs.txt", "r") as f:
    simName = f.read()

with open(data_dir + "simName_obs.txt", "w") as f:
    for i in set(simName.split("\n")):
        if i != "":
            f.write(i + "\n")

# 1-(count/(count1*count2))

## target
with open(data_dir + "same_player_target.txt", "w") as f:
    tmp = pd.DataFrame(birth_df.groupby("birth_block")["player_id"].apply(list))
    for id_list in tmp["player_id"]:
        for idx1 in id_list:
            for idx2 in id_list:
                f.write(str(int(idx1)) + "\t" + str(int(idx2)) + "\n")
                
with open(data_dir + "same_player_target.txt", "r") as f:
    data = f.read()

with open(data_dir + "same_player_target.txt", "w") as f:
    for i in set(data.split("\n")):
        if i != "":
            f.write(i + "\n")



# tmp = world_player[["name","player_url"]]
world_player["player_id"] = world_player["player_url"].apply(lambda x : player2id[x])
who_player["player_id"] = who_player["player_url"].apply(lambda x : player2id[x])


## playerName_obs.txt
with open("PSL_data/playerName_obs.txt", "w") as f:
    for idx, player in id2player.items():
        f.write(str(idx) + "\t" + player + "\n")
