==> KG_generation/Entity Resolution/3. Team Entity Resolution.ipynb <==
!pip install strsimpy
from strsimpy.jaro_winkler import JaroWinkler
from strsimpy.normalized_levenshtein import NormalizedLevenshtein

import pandas as pd
import numpy as np
import json
import fileinput
from tqdm import tqdm
import string
import re
import json
from os import walk
import os
import ast
import re

data_dir = "/Users/jaeyoungkim/Desktop/usc/DSCI - 558/project/sub/structured_data/"

wiki_team = pd.DataFrame()
with fileinput.input(files="../" + "wikipedia_crawler/team_info_wikipedia.jl") as file:
    for line in file:
        conv = json.loads(line)
        wiki_team = wiki_team.append(conv, ignore_index=True)
wiki_team.head()

world_team = pd.DataFrame()
with fileinput.input(files="../" + "worldfootball_crawler/team_info_worldfootball.jl") as file:
    for line in file:
        conv = json.loads(line)
        world_team = world_team.append(conv, ignore_index=True)
world_team.head()

who_team = pd.DataFrame()
with fileinput.input(files="../" + "whoscored_crawler/team_whoscored.jl") as file:
    for line in file:
        conv = json.loads(line)
        who_team = who_team.append(conv, ignore_index=True)
who_team.head()

wiki_team["Fullname"] = wiki_team["Fullname"].apply(lambda  x : x[0])

def f(x):
    for idx1, item in enumerate(x):
        for idx, month in enumerate(["january", "february", "march", "april", "may", "june", "july", "august", "september", "october", "november", "december"]):
            x[idx1] = re.sub(month, "-" + str(idx+1) + "-", x[idx1]).split(";")[0]
    r = re.compile("\d{1,4}-\d{1,2}-\d{1,4}|\d{4}$|\d{2}-\d{2}-\d{4}|\d{2}-\d{1}-\d{4}|\d{2}-\d{4}")
    filtered_list = list(filter(r.match, x))
    return filtered_list

wiki_team['Founded'] = wiki_team["Founded"].apply(lambda x : set(f([i.lower() for i in x])))

Lev = NormalizedLevenshtein().similarity

wiki_name2idx = dict()
wiki_idx2name = dict()
for idx, name in enumerate(wiki_team["name"]):
    wiki_name2idx[name] = idx
    wiki_idx2name[idx] = name

def f(x):
    tmp = [0,0]
    for name in wiki_team["name"]:
        if Lev(x, name) > 0.3 and Lev(x, name) > tmp[1]:
            tmp[0] = wiki_name2idx[name]
            tmp[1] = Lev(x,name)
    if tmp != [0,0]:
        return tmp[0]

who_team["id"] = who_team["name"].apply(f)
world_team["id"] = world_team["name"].apply(f)


wiki_team["id"] = wiki_team["name"].apply(lambda x : wiki_name2idx[x])

who_team.loc[who_team.name == "Bolton", "id"] = 18

who_team.loc[who_team.name == "Brighton", "id"] = 35

team_merged = wiki_team.merge(world_team, left_on = "id", right_on = 'id', how = "outer")
team_merged = team_merged.merge(who_team[["name", "team_url", "id"]].drop_duplicates().rename({"team_url" : "who_team_url"}, axis = 1), left_on = "id", right_on = 'id', how = "outer")

team_merged.loc[38,"name"] = "Brighton"
team_merged.loc[38, "who_team_url"] = 'https://www.whoscored.com/Teams/211/Show/England-Brighton'
team_merged = team_merged.drop(35, axis = 0)
team_merged.to_csv("../data_/" + "team_join.csv")

match = team_merged[["wiki_team_url", "world_team_url", "who_team_url"]]
match = match.append({"wiki_team_url" : 'https://en.wikipedia.org/wiki/Brentford_F.C.', "world_team_url" : 'https://www.worldfootball.net/teams/brentford-fc/', "who_team_url" : 'https://www.whoscored.com/Teams/189/Show/England-Brentford'}, ignore_index = True)
match["id"] = [i for i in range(len(match))]

# match.loc[38,"who_team_url"] = 'https://www.whoscored.com/Teams/211/Show/England-Brighton'
# match = match.drop(35, axis = 0)

match.to_csv("../data_/" + "team_match.csv", index = None)
