==> KG_generation/Entity Resolution/4. Game Entity Resolution.ipynb <==
!pip install strsimpy
from strsimpy.jaro_winkler import JaroWinkler
from strsimpy.normalized_levenshtein import NormalizedLevenshtein

import pandas as pd
import numpy as np
import json
import fileinput
from tqdm import tqdm
import string
import re
import json
from os import walk
import os
import ast
import re

data_dir = "../data_/structured_data/"

f = []
for (dirpath, dirnames, filenames) in walk(data_dir):
    f.extend(filenames)

tot_game_df = pd.DataFrame()
for file in f:
    if file[0] == "s":
        df = pd.read_csv(data_dir + file)
        tot_game_df = tot_game_df.append(df)

tot_game_df

def f(x):
    if "/" in x:
        tmp = x.split("/")
        tmp = tmp[2] + "-" + tmp[1] + "-" + tmp[0]
        if len(tmp) == 8:
            return "20" + tmp
        return tmp
    elif len(x) == 8:
        return '20' + x
    else:
        return x
tot_game_df["date"] = tot_game_df["Date"].apply(lambda x : f(x))

data_dir = "../"

world_game = pd.DataFrame()
with fileinput.input(files=data_dir + "worldfootball_crawler/game_worldfootball.jl") as file:
    for line in file:
        conv = json.loads(line)
        world_game = world_game.append(conv, ignore_index=True)
world_game.head()

len(world_game)

who_game = pd.DataFrame()
with fileinput.input(files=data_dir + "whoscored_crawler/games_whoscored.jl") as file:
    for line in file:
        conv = json.loads(line)
        who_game = who_game.append(conv, ignore_index=True)
who_game.head()

len(who_game)

month_dict = {"Jan" : "01", "Feb" : "02", "Mar" : "03", "Apr" : "04", "May" : "05", "Jun" : "06", "Jul" : "07", "Aug" : "08", "Sep" : "09", "Oct" : "10", "Nov" : "11", "Dec" : 12}

who_game["date"] = who_game["date"].apply(lambda x : x.split(",")[1].split("-")).apply(lambda x : "20" + str(x[2]) + "-" + str(month_dict[x[1]]) + "-" + str(x[0]).strip())


world_game["date"] = world_game["data"].apply(lambda x : x.split("/"))
world_game["date"] = world_game["date"].apply(lambda x : x[2] + "-" + x[1] + "-" + x[0])

## get team match
team_match = pd.read_csv("../data_/" + "team_match.csv", index_col = 'id')

world_game["home_team_id"] = world_game['home_team_url'].apply(lambda x : team_match[team_match["world_team_url"] == x].index[0])

world_game["away_team_id"] = world_game['away_team_url'].apply(lambda x : team_match[team_match["world_team_url"] == x].index[0])




def f(x):
    tmp = x
    if x == 'https://www.whoscored.com/Teams/175/Show/England-West-Brom':
        tmp = 'https://www.whoscored.com/Teams/175/Show/England-West-Bromwich-Albion'

    def g(x):
        if type(x)== float:
            return None
        else:
            return x.split("/")[4]
    
    return team_match[team_match["who_team_url"].apply(lambda x : g(x)) == tmp.split("/")[4]].index[0]
who_game["home_team_id"] = who_game['home_team_url'].apply(lambda x : f(x))
who_game["away_team_id"] = who_game['away_team_url'].apply(lambda x : f(x))


team_match["team_name"] = team_match["wiki_team_url"].apply(lambda x : x.split("/")[-1])

wiki_name2idx = dict()
wiki_idx2name = dict()
for idx, row in team_match.iterrows():
    wiki_name2idx[row.team_name] = idx
    wiki_idx2name[idx] = row.team_name

Lev = NormalizedLevenshtein().similarity

def f(x):
    tmp = [0,0]
    for name in team_match["team_name"]:
        if Lev(x, name) > 0.3 and Lev(x, name) > tmp[1]:
            tmp[0] = wiki_name2idx[name]
            tmp[1] = Lev(x,name)
    if x == "Wolves":
        return 3
    if x == "QPR":
        return 24
    if x == 'Man City':
        return 7
    if tmp != [0,0]:
        return int(tmp[0])

tot_game_df["home_team_id"] = tot_game_df["HomeTeam"].apply(f)
tot_game_df["away_team_id"] = tot_game_df["AwayTeam"].apply(f)

world_who_game = world_game.merge(who_game, left_on = ["home_team_id", "away_team_id", "date"], right_on = ["home_team_id", "away_team_id", "date"], how = "outer", suffixes = ["_world", "_who"])

all_merged = world_who_game.merge(tot_game_df, left_on = ["home_team_id", "away_team_id", "date"], right_on = ["home_team_id", "away_team_id", "date"], how = "outer", suffixes = ["_world_who", "_structured"])


all_merged.to_csv("../data_/" + "game_join.csv", index = None)

len(tot_game_df)

len(world_who_game)

# Div = League Division
# Date = Match Date (dd/mm/yy)
# HomeTeam = Home Team
# AwayTeam = Away Team
# FTHG = Full Time Home Team Goals
# FTAG = Full Time Away Team Goals
# FTR = Full Time Result (H=Home Win, D=Draw, A=Away Win)
# HTHG = Half Time Home Team Goals
# HTAG = Half Time Away Team Goals
# HTR = Half Time Result (H=Home Win, D=Draw, A=Away Win)

# Match Statistics (where available)
# Attendance = Crowd Attendance
# Referee = Match Referee
# HS = Home Team Shots
# AS = Away Team Shots
# HST = Home Team Shots on Target
# AST = Away Team Shots on Target
# HHW = Home Team Hit Woodwork
# AHW = Away Team Hit Woodwork
# HC = Home Team Corners
# AC = Away Team Corners
# HF = Home Team Fouls Committed
# AF = Away Team Fouls Committed
# HO = Home Team Offsides
# AO = Away Team Offsides
# HY = Home Team Yellow Cards
# AY = Away Team Yellow Cards
# HR = Home Team Red Cards
# AR = Away Team Red Cards
# HBP = Home Team Bookings Points (10 = yellow, 25 = red)
# ABP = Away Team Bookings Points (10 = yellow, 25 = red)

# Key to 1X2 (match) betting odds data:

# B365H = Bet365 home win odds
# B365D = Bet365 draw odds
# B365A = Bet365 away win odds
# BSH = Blue Square home win odds
# BSD = Blue Square draw odds
# BSA = Blue Square away win odds
# BWH = Bet&Win home win odds
# BWD = Bet&Win draw odds
# BWA = Bet&Win away win odds
# GBH = Gamebookers home win odds
# GBD = Gamebookers draw odds
# GBA = Gamebookers away win odds
# IWH = Interwetten home win odds
# IWD = Interwetten draw odds
# IWA = Interwetten away win odds
# LBH = Ladbrokes home win odds
# LBD = Ladbrokes draw odds
# LBA = Ladbrokes away win odds
# PSH = Pinnacle Sports home win odds
# PSD = Pinnacle Sports draw odds
# PSA = Pinnacle Sports away win odds
# SOH = Sporting Odds home win odds
# SOD = Sporting Odds draw odds
# SOA = Sporting Odds away win odds
# SBH = Sportingbet home win odds
# SBD = Sportingbet draw odds
# SBA = Sportingbet away win odds
# SJH = Stan James home win odds
# SJD = Stan James draw odds
# SJA = Stan James away win odds
# SYH = Stanleybet home win odds
# SYD = Stanleybet draw odds
# SYA = Stanleybet away win odds
# VCH = VC Bet home win odds
# VCD = VC Bet draw odds
# VCA = VC Bet away win odds
# WHH = William Hill home win odds
# WHD = William Hill draw odds
# WHA = William Hill away win odds

# Bb1X2 = Number of BetBrain bookmakers used to calculate match odds averages and maximums
# BbMxH = Betbrain maximum home win odds
# BbAvH = Betbrain average home win odds
# BbMxD = Betbrain maximum draw odds
# BbAvD = Betbrain average draw win odds
# BbMxA = Betbrain maximum away win odds
# BbAvA = Betbrain average away win odds



# Key to total goals betting odds:

# BbOU = Number of BetBrain bookmakers used to calculate over/under 2.5 goals (total goals) averages and maximums
# BbMx>2.5 = Betbrain maximum over 2.5 goals
# BbAv>2.5 = Betbrain average over 2.5 goals
# BbMx<2.5 = Betbrain maximum under 2.5 goals
# BbAv<2.5 = Betbrain average under 2.5 goals

# GB>2.5 = Gamebookers over 2.5 goals
# GB<2.5 = Gamebookers under 2.5 goals
# B365>2.5 = Bet365 over 2.5 goals
# B365<2.5 = Bet365 under 2.5 goals


# Key to Asian handicap betting odds:

# BbAH = Number of BetBrain bookmakers used to Asian handicap averages and maximums
# BbAHh = Betbrain size of handicap (home team)
# BbMxAHH = Betbrain maximum Asian handicap home team odds
# BbAvAHH = Betbrain average Asian handicap home team odds
# BbMxAHA = Betbrain maximum Asian handicap away team odds
# BbAvAHA = Betbrain average Asian handicap away team odds

# GBAHH = Gamebookers Asian handicap home team odds
# GBAHA = Gamebookers Asian handicap away team odds
# GBAH = Gamebookers size of handicap (home team)
# LBAHH = Ladbrokes Asian handicap home team odds
# LBAHA = Ladbrokes Asian handicap away team odds
# LBAH = Ladbrokes size of handicap (home team)
# B365AHH = Bet365 Asian handicap home team odds
# B365AHA = Bet365 Asian handicap away team odds
# B365AH = Bet365 size of handicap (home team)

# Football-Data would like to acknowledge the following sources which have been utilised in the compilation of Football-Data's results and odds files.

# Historical results:
# International Soccer Server - http://sunsite.tut.fi/rec/riku/soccer.html
# European Football - http://www.eurofootball.be/
# RSSSF Archive - http://www.rsssf.com/

# Current results (full time, half time)
# TBWSport - http://www.tbwsport.com
# Livescore- http://www.livescore.com

# Match statistics
# Sportinglife, ESPN Soccer, Bundesliga.de, Gazzetta.it and Football.fr

# Bookmakers betting odds
# Betbrain - http://www.betbrain.com
# Betbase - http://www.betbase.info

# Betting odds for weekend games are collected Friday afternoons, and on Tuesday afternoons for midweek games.
